+++ src/mkvtoolnix-gui/header_editor/track_name_page.h
@@ -29,6 +29,8 @@ public:
   virtual bool validateValue() const override;
   virtual void copyValueToElement() override;
 
+  virtual void setString(QString const &value);
+
 public Q_SLOTS:
   virtual void setupPredefinedTrackNames();
 };